\name{gut.Ds}
\alias{gut.Ds}
\docType{data}
\title{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\description{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\usage{data("gut.Ds")}
\format{
  The format is:
List of 7
 $ : num [1:23, 1:23] 0 0.719 0.692 0.822 0.745 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.666 0.716 0.766 0.803 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.441 0.501 0.596 0.548 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.535 0.588 0.705 0.733 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.397 0.447 0.577 0.626 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.275 0.323 0.451 0.523 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
 $ : num [1:23, 1:23] 0 0.188 0.237 0.352 0.449 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
  .. ..$ : chr [1:23] "JMat-baseline-99R1062" "JMat-baseline-DCGG" "JMat-baseline-DCJT" "JMat-baseline-DE76A" ...
}
\references{
Yanai H, Park B, Koh H. Jang HJ, Vaughan KL, Tanaka-Yano M, Aon M, Blanton M, Messaoudi I, Diaz-Ruiz A, Mattison JA, Beerman I. Short-term periodic restricted feeding elicits metabolome-microbiome signatures with sex dimorphic persistence in primate intervention. Nat Commun, 15(1088), 2024

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(gut.Ds)
}
\keyword{datasets}
