\name{gut.meta}
\alias{gut.meta}
\docType{data}
\title{
Metadata for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\description{
Metadata for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\usage{data("gut.meta")}
\format{
  A data frame with 23 observations on the following 4 variables.
  \describe{
    \item{\code{body.weight}}{a numeric vector}
    \item{\code{diet}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
  }
}
\references{
Yanai H, Park B, Koh H. Jang HJ, Vaughan KL, Tanaka-Yano M, Aon M, Blanton M, Messaoudi I, Diaz-Ruiz A, Mattison JA, Beerman I. Short-term periodic restricted feeding elicits metabolome-microbiome signatures with sex dimorphic persistence in primate intervention. Nat Commun, 15(1088), 2024

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(gut.meta)
}
\keyword{datasets}
