\name{gut.otu.tab}
\alias{gut.otu.tab}
\docType{data}
\title{
A feature table for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\description{
A feature table for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\usage{data("gut.otu.tab")}
\format{
  A data frame with 419 observations on the following 23 variables.
  \describe{
    \item{\samp{JMat-baseline-99R1062}}{a numeric vector}
    \item{\samp{JMat-baseline-DCGG}}{a numeric vector}
    \item{\samp{JMat-baseline-DCJT}}{a numeric vector}
    \item{\samp{JMat-baseline-DE76A}}{a numeric vector}
    \item{\samp{JMat-baseline-DF11}}{a numeric vector}
    \item{\samp{JMat-baseline-DF46}}{a numeric vector}
    \item{\samp{JMat-baseline-M1352}}{a numeric vector}
    \item{\samp{JMat-baseline-M1423}}{a numeric vector}
    \item{\samp{JMat-baseline-MLB}}{a numeric vector}
    \item{\samp{JMat-baseline-MRI}}{a numeric vector}
    \item{\samp{JMat-baseline-MTR}}{a numeric vector}
    \item{\samp{JMat-baseline-RQ2937}}{a numeric vector}
    \item{\samp{JMat-baseline-RQ6683}}{a numeric vector}
    \item{\samp{JMat-baseline-RQ6709}}{a numeric vector}
    \item{\samp{JMat-baseline-T19}}{a numeric vector}
    \item{\samp{JMat-baseline-T75}}{a numeric vector}
    \item{\samp{JMat-baseline-ZA49}}{a numeric vector}
    \item{\samp{JMat-baseline-ZA64}}{a numeric vector}
    \item{\samp{JMat-baseline-ZE26}}{a numeric vector}
    \item{\samp{JMat-baseline-ZF36}}{a numeric vector}
    \item{\samp{JMat-baseline-ZF77}}{a numeric vector}
    \item{\samp{JMat-baseline-ZJ51}}{a numeric vector}
    \item{\samp{JMat-baseline-ZJ55}}{a numeric vector}
  }
}
\references{
Yanai H, Park B, Koh H. Jang HJ, Vaughan KL, Tanaka-Yano M, Aon M, Blanton M, Messaoudi I, Diaz-Ruiz A, Mattison JA, Beerman I. Short-term periodic restricted feeding elicits metabolome-microbiome signatures with sex dimorphic persistence in primate intervention. Nat Commun, 15(1088), 2024

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(gut.otu.tab)
}
\keyword{datasets}
