\name{OmniK-package}
\alias{OmniK-package}
\alias{OmniK}
\docType{package}
\title{
\packageTitle{OmniK}
}
\description{
\packageDescription{OmniK}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{OmniK}

Maintainer: \packageMaintainer{OmniK}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = OmniK)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{OmniK}}
%\section{Documentation Index}{\packageIndices{OmniK}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
