\name{oral.Ds}
\alias{oral.Ds}
\docType{data}
\title{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\description{
A list of ecological (pairwise, subject-by-subject, n by n) distance matrices for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\usage{data("oral.Ds")}
\format{
  The format is:
List of 7
 $ : num [1:145, 1:145] 0 0.523 0.525 0.531 0.548 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.497 0.476 0.465 0.546 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.173 0.167 0.162 0.196 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.297 0.348 0.325 0.37 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.195 0.251 0.223 0.259 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.137 0.185 0.161 0.191 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
 $ : num [1:145, 1:145] 0 0.101 0.139 0.119 0.146 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
  .. ..$ : chr [1:145] "V100" "V101" "V102" "V103" ...
}
\references{
Park B, Koh H, Patatanian M, Reyes-Caballero H, Zhao N, Meinert J, Holbrook JT, Leinbach LI, Biswal S. The mediating roles of the oral microbiome in saliva and subgingival sites between e-cigarette smoking and gingival inflammation. BMC Microbiol, 23(35), 2023

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(oral.Ds)
}
\keyword{datasets}
