\name{oral.meta}
\alias{oral.meta}
\docType{data}
\title{
Metadata for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\description{
Metadata for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\usage{data("oral.meta")}
\format{
  A data frame with 145 observations on the following 4 variables.
  \describe{
    \item{\code{gingival.inflammation}}{a numeric vector}
    \item{\code{ecig.use}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
  }
}
\references{
Park B, Koh H, Patatanian M, Reyes-Caballero H, Zhao N, Meinert J, Holbrook JT, Leinbach LI, Biswal S. The mediating roles of the oral microbiome in saliva and subgingival sites between e-cigarette smoking and gingival inflammation. BMC Microbiol, 23(35), 2023

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(oral.meta)
}
\keyword{datasets}
