\name{oral.otu.tab}
\alias{oral.otu.tab}
\docType{data}
\title{
A feature table for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\description{
A feature table for the oral microbiome data (Park et al., 2023, BMC Microbiology)
}
\usage{data("oral.otu.tab")}
\format{
  A data frame with 1547 observations on the following 145 variables.
  \describe{
    \item{\code{V100}}{a numeric vector}
    \item{\code{V101}}{a numeric vector}
    \item{\code{V102}}{a numeric vector}
    \item{\code{V103}}{a numeric vector}
    \item{\code{V104}}{a numeric vector}
    \item{\code{V105}}{a numeric vector}
    \item{\code{V108}}{a numeric vector}
    \item{\code{V109}}{a numeric vector}
    \item{\code{V110}}{a numeric vector}
    \item{\code{V112}}{a numeric vector}
    \item{\code{V113}}{a numeric vector}
    \item{\code{V114}}{a numeric vector}
    \item{\code{V116}}{a numeric vector}
    \item{\code{V117}}{a numeric vector}
    \item{\code{V118}}{a numeric vector}
    \item{\code{V120}}{a numeric vector}
    \item{\code{V122}}{a numeric vector}
    \item{\code{V123}}{a numeric vector}
    \item{\code{V126}}{a numeric vector}
    \item{\code{V130}}{a numeric vector}
    \item{\code{V133}}{a numeric vector}
    \item{\code{V134}}{a numeric vector}
    \item{\code{V135}}{a numeric vector}
    \item{\code{V137}}{a numeric vector}
    \item{\code{V139}}{a numeric vector}
    \item{\code{V148}}{a numeric vector}
    \item{\code{V150}}{a numeric vector}
    \item{\code{V151}}{a numeric vector}
    \item{\code{V154}}{a numeric vector}
    \item{\code{V155}}{a numeric vector}
    \item{\code{V157}}{a numeric vector}
    \item{\code{V162}}{a numeric vector}
    \item{\code{V164}}{a numeric vector}
    \item{\code{V166}}{a numeric vector}
    \item{\code{V167}}{a numeric vector}
    \item{\code{V168}}{a numeric vector}
    \item{\code{V169}}{a numeric vector}
    \item{\code{V170}}{a numeric vector}
    \item{\code{V171}}{a numeric vector}
    \item{\code{V172}}{a numeric vector}
    \item{\code{V174}}{a numeric vector}
    \item{\code{V175}}{a numeric vector}
    \item{\code{V176}}{a numeric vector}
    \item{\code{V180}}{a numeric vector}
    \item{\code{V181}}{a numeric vector}
    \item{\code{V184}}{a numeric vector}
    \item{\code{V186}}{a numeric vector}
    \item{\code{V188}}{a numeric vector}
    \item{\code{V191}}{a numeric vector}
    \item{\code{V192}}{a numeric vector}
    \item{\code{V193}}{a numeric vector}
    \item{\code{V194}}{a numeric vector}
    \item{\code{V195}}{a numeric vector}
    \item{\code{V196}}{a numeric vector}
    \item{\code{V197}}{a numeric vector}
    \item{\code{V198}}{a numeric vector}
    \item{\code{V199}}{a numeric vector}
    \item{\code{V200}}{a numeric vector}
    \item{\code{V201}}{a numeric vector}
    \item{\code{V202}}{a numeric vector}
    \item{\code{V203}}{a numeric vector}
    \item{\code{V204}}{a numeric vector}
    \item{\code{V209}}{a numeric vector}
    \item{\code{V212}}{a numeric vector}
    \item{\code{V215}}{a numeric vector}
    \item{\code{V216}}{a numeric vector}
    \item{\code{V217}}{a numeric vector}
    \item{\code{V218}}{a numeric vector}
    \item{\code{V219}}{a numeric vector}
    \item{\code{V222}}{a numeric vector}
    \item{\code{V223}}{a numeric vector}
    \item{\code{V224}}{a numeric vector}
    \item{\code{V225}}{a numeric vector}
    \item{\code{V227}}{a numeric vector}
    \item{\code{V228}}{a numeric vector}
    \item{\code{V233}}{a numeric vector}
    \item{\code{V234}}{a numeric vector}
    \item{\code{V235}}{a numeric vector}
    \item{\code{V236}}{a numeric vector}
    \item{\code{V237}}{a numeric vector}
    \item{\code{V238}}{a numeric vector}
    \item{\code{V240}}{a numeric vector}
    \item{\code{V241}}{a numeric vector}
    \item{\code{V248}}{a numeric vector}
    \item{\code{V250}}{a numeric vector}
    \item{\code{V251}}{a numeric vector}
    \item{\code{V253}}{a numeric vector}
    \item{\code{V256}}{a numeric vector}
    \item{\code{V257}}{a numeric vector}
    \item{\code{V258}}{a numeric vector}
    \item{\code{V259}}{a numeric vector}
    \item{\code{V262}}{a numeric vector}
    \item{\code{V263}}{a numeric vector}
    \item{\code{V265}}{a numeric vector}
    \item{\code{V266}}{a numeric vector}
    \item{\code{V273}}{a numeric vector}
    \item{\code{V275}}{a numeric vector}
    \item{\code{V276}}{a numeric vector}
    \item{\code{V278}}{a numeric vector}
    \item{\code{V282}}{a numeric vector}
    \item{\code{V285}}{a numeric vector}
    \item{\code{V289}}{a numeric vector}
    \item{\code{V293}}{a numeric vector}
    \item{\code{V295}}{a numeric vector}
    \item{\code{V297}}{a numeric vector}
    \item{\code{V298}}{a numeric vector}
    \item{\code{V302}}{a numeric vector}
    \item{\code{V305}}{a numeric vector}
    \item{\code{V306}}{a numeric vector}
    \item{\code{V310}}{a numeric vector}
    \item{\code{V311}}{a numeric vector}
    \item{\code{V314}}{a numeric vector}
    \item{\code{V315}}{a numeric vector}
    \item{\code{V316}}{a numeric vector}
    \item{\code{V319}}{a numeric vector}
    \item{\code{V321}}{a numeric vector}
    \item{\code{V325}}{a numeric vector}
    \item{\code{V326}}{a numeric vector}
    \item{\code{V327}}{a numeric vector}
    \item{\code{V328}}{a numeric vector}
    \item{\code{V329}}{a numeric vector}
    \item{\code{V330}}{a numeric vector}
    \item{\code{V331}}{a numeric vector}
    \item{\code{V334}}{a numeric vector}
    \item{\code{V335}}{a numeric vector}
    \item{\code{V339}}{a numeric vector}
    \item{\code{V340}}{a numeric vector}
    \item{\code{V342}}{a numeric vector}
    \item{\code{V343}}{a numeric vector}
    \item{\code{V344}}{a numeric vector}
    \item{\code{V346}}{a numeric vector}
    \item{\code{V347}}{a numeric vector}
    \item{\code{V352}}{a numeric vector}
    \item{\code{V353}}{a numeric vector}
    \item{\code{V354}}{a numeric vector}
    \item{\code{V356}}{a numeric vector}
    \item{\code{V358}}{a numeric vector}
    \item{\code{V359}}{a numeric vector}
    \item{\code{V363}}{a numeric vector}
    \item{\code{V368}}{a numeric vector}
    \item{\code{V369}}{a numeric vector}
    \item{\code{V371}}{a numeric vector}
    \item{\code{V372}}{a numeric vector}
    \item{\code{V375}}{a numeric vector}
    \item{\code{V380}}{a numeric vector}
  }
}
\references{
Park B, Koh H, Patatanian M, Reyes-Caballero H, Zhao N, Meinert J, Holbrook JT, Leinbach LI, Biswal S. The mediating roles of the oral microbiome in saliva and subgingival sites between e-cigarette smoking and gingival inflammation. BMC Microbiol, 23(35), 2023

Koh H. A general kernel machine regression framework using principal component analysis for jointly testing main and interaction effects. (In Review)
}
\examples{
data(oral.otu.tab)
}
\keyword{datasets}
